+++ Sources/MapboxMaps/Style/Style.swift
@@ -67,16 +67,15 @@ public class Style {
 
     /**
      Gets a `layer` from the map
-     - Parameter layerID: The id of the layer to be fetched
+     - Parameter id: The id of the layer to be fetched
      - Parameter type: The type of the layer that will be fetched
 
      - Returns: The fully formed `layer` object of type equal to `type`
      - Throws: StyleError or type conversion errors
      */
-    public func layer<T: Layer>(withId id: String) throws -> T {
-        // swiftlint:disable force_cast
-        return try layer(withId: id, type: T.self) as! T
-        // swiftlint:enable force_cast
+    public func layer<T>(withId id: String, type: T.Type) throws -> T where T: Layer {
+        let properties = try layerProperties(for: id)
+        return try type.init(jsonObject: properties)
     }
 
     /**
@@ -86,17 +85,20 @@ public class Style {
      you are fetching, or don't need to know for your situation.
 
      - Parameter layerID: The id of the layer to be fetched
-     - Parameter type: The type of the layer that will be fetched
 
-     - Returns: The fully formed `layer` object of type equal to `type`
-     - Throws: StyleError or type conversion errors
-     - Note: This method is marked as experimental. Annotate the import statement
-     for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+     - Returns: The fully formed `layer` object.
+     - Throws: Type conversion errors
      */
-    @_spi(Experimental) public func layer(withId id: String, type: Layer.Type) throws -> Layer {
+    public func layer(withId id: String) throws -> Layer {
         // Get the layer properties from the map
         let properties = try layerProperties(for: id)
-        return try type.init(jsonObject: properties)
+
+        guard let typeString = properties["type"] as? String,
+              let type = LayerType(rawValue: typeString) else {
+            throw TypeConversionError.invalidObject
+        }
+
+        return try type.layerType.init(jsonObject: properties)
     }
 
     /// Updates a layer that exists in the style already
@@ -107,12 +109,13 @@ public class Style {
     ///   - update: Closure that mutates a layer passed to it
     ///
     /// - Throws: StyleError or type conversion errors
-    public func updateLayer<T: Layer>(withId id: String, update: (inout T) throws -> Void) throws {
-        var layer: T = try self.layer(withId: id)
+    public func updateLayer<T>(withId id: String,
+                               type: T.Type,
+                               update: (inout T) throws -> Void) throws where T: Layer {
+        var layer = try self.layer(withId: id, type: T.self)
 
         // Call closure to update the retrieved layer
         try update(&layer)
-
         let value = try layer.jsonObject()
 
         // Apply the changes to the layer properties to the style
@@ -149,15 +152,15 @@ public class Style {
 
     /**
      Retrieves a source from the map
-     - Parameter identifier: The id of the source to retrieve
+     - Parameter id: The id of the source to retrieve
      - Parameter type: The type of the source
+
      - Returns: The fully formed `source` object of type equal to `type`.
      - Throws: StyleError or type conversion errors
      */
-    public func source<T: Source>(withId id: String) throws -> T {
-        // swiftlint:disable force_cast
-        return try source(withId: id, type: T.self) as! T
-        // swiftlint:enable force_cast
+    public func source<T>(withId id: String, type: T.Type) throws -> T where T: Source {
+        let sourceProps = try sourceProperties(for: id)
+        return try type.init(jsonObject: sourceProps)
     }
 
     /**
@@ -166,18 +169,19 @@ public class Style {
      This function is useful if you do not know the concrete type of the source
      you are fetching, or don't need to know for your situation.
 
-     - Parameter identifier: The id of the source to retrieve
-     - Parameter type: The type of the source
-     - Returns: The fully formed `source` object of type equal to `type`.
-     - Throws: StyleError or type conversion errors
-     - Note: This method is marked as experimental. Annotate the import statement
-     for `MapboxMaps` with `@_spi(Experimental)` in order to use experimental methods.
+     - Parameter id: The id of the source to retrieve.
+     - Returns: The fully formed `source` object.
+     - Throws: Type conversion errors.
      */
-    @_spi(Experimental) public func source(withId id: String, type: Source.Type) throws  -> Source {
+    public func source(withId id: String) throws  -> Source {
         // Get the source properties for a given identifier
         let sourceProps = try sourceProperties(for: id)
-        let source = try type.init(jsonObject: sourceProps)
-        return source
+
+        guard let typeString = sourceProps["type"] as? String,
+              let type = SourceType(rawValue: typeString) else {
+            throw TypeConversionError.invalidObject
+        }
+        return try type.sourceType.init(jsonObject: sourceProps)
     }
 
     /// Updates the `data` property of a given `GeoJSONSource` with a new value