+++ Tests/MapboxMapsTests/Gestures/GestureHandlers/Mocks/MockGestureHandlerDelegate.swift
@@ -5,4 +5,18 @@ final class MockGestureHandlerDelegate: GestureHandlerDelegate {
     func gestureBegan(for gestureType: GestureType) {
         gestureBeganStub.call(with: gestureType)
     }
+
+    struct GestureEndedParams {
+        var gestureType: GestureType
+        var willAnimate: Bool
+    }
+    let gestureEndedStub = Stub<GestureEndedParams, Void>()
+    func gestureEnded(for gestureType: GestureType, willAnimate: Bool) {
+        gestureEndedStub.call(with: GestureEndedParams(gestureType: gestureType, willAnimate: willAnimate))
+    }
+
+    let animationEndedStub = Stub<GestureType, Void>()
+    func animationEnded(for gestureType: GestureType) {
+        animationEndedStub.call(with: gestureType)
+    }
 }